% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_to_factor.R
\name{change_to_factor}
\alias{change_to_factor}
\title{Convert character variables in formula to factor variables.
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.}
\usage{
change_to_factor(df, vars)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{A vector of dataframe column names}
}
\value{
Dataframe where all character variables are now factors
}
\description{
Convert character variables in formula to factor variables.
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.
}
\examples{
\dontrun{
df <- data.frame(x = 1:3, y = 3:1, z = c("a", "b", "c"),
                 stringsAsFactors = FALSE)
fm <- as.formula(x ~ y + z)
vars <- all.vars(fm)
new_df <- change_to_factor(df, vars)
lapply(new_df, class)
}

}
\keyword{internal}
