% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{check_lookback}
\alias{check_lookback}
\title{Run checks on variable lookback}
\usage{
check_lookback(data, lookback, entry)
}
\arguments{
\item{lookback}{See rollmatch()}
}
\description{
Run checks on variable lookback
}
\keyword{internal}
