% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{chr_2_factor}
\alias{chr_2_factor}
\title{Convert character variables in formula to factor variables.
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.}
\usage{
chr_2_factor(df, vars)
}
\arguments{
\item{df}{A dataframe.}

\item{vars}{A vector of dataframe column names}
}
\value{
Dataframe where all character variables are now factors
}
\description{
Convert character variables in formula to factor variables.
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.
}
\examples{
\dontrun{
df <- data.frame(x = 1:3, y = 3:1, z = c("a", "b", "c"),
                 stringsAsFactors = FALSE)
vars <- names(df)
newdf <- df <- chr_2_factor(df, vars)
lapply(df, class)
}

}
\keyword{internal}
