% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_comparison}
\alias{create_comparison}
\title{Create a dataframe of comparisons between all treatment and control data.
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.}
\usage{
create_comparison(scored_data, treatm, tm, entry, id)
}
\arguments{
\item{scored_data}{The dataset given to runModel with the additional of the
score values generated by the propensity score model}

\item{tm}{The time period indicator}

\item{entry}{Entry quarter for subject}

\item{id}{ID variable in dataset}
}
\value{
Dataframe comparing all treatment and control data
}
\description{
Create a dataframe of comparisons between all treatment and control data.
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.
}
\examples{
\dontrun{ 
# As these functions are hidden, we only provided one example.
# See add_balance_table function in the utility_functions file.
}

}
\keyword{internal}
