% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_weights}
\alias{create_weights}
\title{Create control weights for matches dataset, and final data output
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.}
\usage{
create_weights(matches, data, id)
}
\arguments{
\item{matches}{Dataframe containing the matches from comparison_pool}

\item{data}{The original data provided for the function.}

\item{id}{The individual id variable.}
}
\value{
A list containing two Dataframes. Matches - an updated dataset with
control weights added, and data_full an updated version of the original data
with weights added.
}
\description{
Create control weights for matches dataset, and final data output
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.
}
\examples{
\dontrun{ 
# As these functions are hidden, we only provided one example.
# See add_balance_table function in the utility_functions file.
}

}
\keyword{internal}
