% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output.R
\name{make_output}
\alias{make_output}
\title{Combine the results of rollmatch into a tidy list for output
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.}
\usage{
make_output(scored_data, data_full, matches, treat, tm, entry, lookback)
}
\arguments{
\item{scored_data}{The dataset given to runModel with the additional of the
score values generated by the propensity score model}

\item{data_full}{The original data provided with the additional of
control weights}

\item{matches}{Dataframe containing the matches from comparison_pool}

\item{tm}{The time period indicator.}

\item{entry}{The time period in which the participant enrolled in the
intervention (in the same units as the tm variable).}

\item{lookback}{The number of time periods to look back before the
time period of enrollment (1-10).}
}
\value{
\code{output} returns a list containing the following components:
\item{scores}{The propensity scores used in matching.}
\item{data}{The original dataset with matches, scores, and weights applied.}
\item{summary}{A basic summary table.}
\item{ids_not_matched}{A vector of unmatched treatment ids}
\item{total_not_matched}{The count of unmatched treatment ids}
}
\description{
Combine the results of rollmatch into a tidy list for output
As it's an internal helper function to aid in testing, it is not exported for use outside of the package.
}
\examples{
\dontrun{ 
formula <- as.formula(treat ~ qtr_pmt + yr_pmt + age + is_male + is_white +
                       is_disabled + is_esrd + months_dual + chron_num + lq_ed +
                       yr_ed2 + lq_ip + yr_ip2)
tm <- "quarter"; entry <- "entry_q"; lookback <- 1
load(url(paste0("https://github.com/RTIInternational/rollmatch/raw/master/",
                "tests/testthat/output.rda")))
scored_data <- output$scored_data
load(url(paste0("https://github.com/RTIInternational/rollmatch/raw/master/",
                "tests/testthat/out_list.rda")))
data_full <- out_list$data_full
matches <- out_list$matches
out <- make_output(scored_data, data_full, matches, tm, entry, lookback)
head(out)
}

}
\keyword{internal}
