% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data.R
\name{reduce_data}
\alias{reduce_data}
\title{Preprocessing Step to Rolling Entry Matching}
\usage{
reduce_data(data, treat, tm, entry, id, lookback = 1)
}
\arguments{
\item{data}{The input panel dataset.}

\item{treat}{The treatment indentifier (should contain 0's and 1's)}

\item{tm}{The time period indicator.}

\item{entry}{The time period in which the participant enrolled in the
intervention (in the same units as the tm variable).}

\item{id}{The individual id variable. This is required because reduced_data
must have an id column}

\item{lookback}{The number of time periods to look back before the
time period of enrollment (1-10).}
}
\value{
\code{reduced_data} returns a dataset of reduced data ready
for propensity scoring
}
\description{
Preprocessing Step to Rolling Entry Matching
}
\examples{
data(package="rollmatch", "rem_synthdata_small")
reduced_data <- reduce_data(data = rem_synthdata_small, treat = "treat",
                            tm = "quarter", entry = "entry_q", id = "indiv_id",
                            lookback = 1)
reduced_data

}
