% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollmatch.R
\name{rollmatch}
\alias{rollmatch}
\title{Rolling entry matching}
\usage{
rollmatch(scored_data, data_full, treat, tm, entry, id, vars, alpha = 0,
  weighted_pooled_stdev = FALSE, num_matches = 3, replacement = TRUE)
}
\arguments{
\item{scored_data}{Input data with propensity scores labeled "score"}

\item{data_full}{The original dataset before reduce_data() was ran}

\item{tm}{The time period indicator.}

\item{entry}{The time period in which the participant enrolled in the
intervention (in the same units as the tm variable).}

\item{id}{The individual id variable.}

\item{vars}{The column names used in the propensity score algorithm}

\item{alpha}{Part of the pre-specified distance within which to allow
matching. The caliper width is calculated as the \code{alpha} multiplied by the
pooled standard deviation of the propensity scores or the logit of the
propensity scores - depending on the value of \code{match_on}.}

\item{weighted_pooled_stdev}{Option that allows for weighted pooled standard
deviation for caliper matching. FALSE for average pooled standard
deviation and TRUE for weighted pooled standard deviation.}

\item{num_matches}{The number of comparison beneficiary matches to attempt
to assign to each treatment beneficiary}

\item{replacement}{Assign comparison beneficiaries with replacement (TRUE)
or without replacement (FALSE). If \code{replacement} is TRUE, then
comparison beneficiaries will be allowed to be used with replacement within
a single quarter, but will not be allowed to match to different treatment
beneficiaries across multiple quarters.}

\item{match_on}{Match on estimated propensity score ("pscore") or logit of
estimated propensity score ("logit").}

\item{model_type}{Use logistic regression ("logistic") or "probit"
regression ("probit") to estimate the predicted probability of participating
in the innovation.}
}
\value{
\code{rollmatch} returns an object of class "rollmatch".

An object of class "rollmatch" is a list containing the following components:
\item{model}{The output of the model used to estimate the distance measure.}
\item{scores}{The propensity score and logit of the propensity score. }
\item{data}{The original dataset with matches, scores, and weights applied.}
\item{summary}{A basic summary table.}
\item{balance}{The balancing table.}
}
\description{
\code{rollmatch} is the main function of the rollmatch package
\emph{<rollmatch>}, which implements a comparison group selection
methodology for interventions with rolling participant entry over time.
A difficulty in evaluating rolling entry interventions is that a suitable
"entry" date is not observed for non-participants. This method, called
rolling entry matching, assigns potential comparison non-participants
multiple counterfactual entry periods which allows for matching of
participant and non-participants based on data immediately preceding each
participant's specific entry period, rather than using data from a fixed
pre-intervention period.
}
\details{
Rolling entry matching requires three preliminary steps. This package will 
assist the user in steps 2 and three. First, a quasi-panel dataset
is constructed containing multiple observations of non-participants (one for
each entry period). Participants enter the data once in the baseline period
immediately preceding their unique entry into the intervention. Time-varying
covariates (e.g., health conditions, spending, utilization) are dynamic for
each entry period's non-participant observations. The user of rollmatch
is expected to have already created this quasi-panel dataset. Second, the
pool of potential comparisons for each participant is restricted to those
that have the same "entry period" into the intervention
(see function "reduceData"). Finally, a predicted
probability of treatment is obtained for participants and non-participants
(e.g. through propensity score matching). The user can use function
"calcScores" to complete this step, or create use their own propensity score 
calculation.

The final step consists of the matching algorithm. The algorithm selects
the best matched comparison(s) for each participant from the pool of
non-participants with the same entry period. This is completed via the
function "rollmatch".
}
\examples{
data(package="rollmatch", "rem_synthdata_small")

reduced_data <- reduce_data(data = rem_synthdata_small, treat = "treat",
                            tm = "quarter", entry = "entry_q",
                            id = "indiv_id", lookback = 1)
                            
                            
formula <- as.formula(treat ~ qtr_pmt + yr_pmt + age)

r_match <- rollmatch(formula, data = rem_synthdata_small, treat = "treat",
                       tm = "quarter", entry = "entry_q", id = "indiv_id",
                       alpha = 0.2)

r_match

}
