% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks_one}
\alias{run_checks_one}
\title{Run checks on variable inputs 
As it's an internal helper function to aid in testing,
it is not exported for use outside of the package.}
\usage{
run_checks_one(data, tm, entry, id)
}
\arguments{
\item{data}{dataset whose columns match the other parameters}

\item{tm}{The time period indicator}

\item{entry}{The time period in which the participant enrolled in the
intervention (in the same units as the tm variable).}

\item{id}{The individual id variable.}
}
\description{
Run checks on variable inputs 
As it's an internal helper function to aid in testing,
it is not exported for use outside of the package.
}
\keyword{internal}
