% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trim_pool}
\alias{trim_pool}
\title{Use a caliper to trim the data to only observations within threshold}
\usage{
trim_pool(alpha, data_pool, scored_data, treat,
  weighted_pooled_stdev = FALSE)
}
\arguments{
\item{alpha}{The pre-specified distance within which to allow matching.
The caliper width is calculated as the \code{alpha} multiplied by the
pooled standard deviation of the propensity scores or the logit of the
propensity scores - depending on the value of \code{match_on}.}

\item{data_pool}{Dataframe of comparison data to be trimmed.}

\item{scored_data}{Dataframe of results from model in runModel.}

\item{weighted_pooled_stdev}{Boolean. FALSE for average pooled standard
deviation and TRUE for weighted pooled standard deviation.}
}
\value{
Dataframe of the trimmed comparisons based on the alpha value
}
\description{
Use a caliper to trim the data to only observations within threshold
}
\examples{
\dontrun{ 
# As these functions are hidden, we only provided one example.
# See add_balance_table function in the utility_functions file.
}
                         
}
\keyword{internal}
