% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_matches_columns}
\alias{add_matches_columns}
\title{Create Additional Columns for the Matches Dataset}
\usage{
add_matches_columns(matches)
}
\arguments{
\item{matches}{Dataframe containing the matches from comparison_pool. Each
row represents a match, and there should be columns for `treat_id`
and possibly `control_id` if control matches are to be calculated.}
}
\value{
A dataframe containing the original match information along with
        additional columns: `match_rank`, `total_matches`, `treatment_weight`,
        `control_matches`, and `control_weight`.
}
\description{
This function takes a dataframe containing match information and adds
additional columns to indicate the match rank, total matches for a given
treatment ID, treatment weight, control matches, and control weight.
}
\examples{
\donttest{ 
print('See add_balance_table for full example')
}

}
\keyword{internal}
