% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{add_matches_columns}
\alias{add_matches_columns}
\title{Create additional columns for the matches dataset}
\usage{
add_matches_columns(matches)
}
\arguments{
\item{Matches}{Dataframe containing the matches from comparison_pool}
}
\value{
Dataframe containing top matches
}
\description{
Create additional columns for the matches dataset
}
\examples{
\dontrun{ 
print('See add_balance_table for full example')
}

}
\keyword{internal}
