% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{addMatchesColumns}
\alias{addMatchesColumns}
\title{Create additional columns for the matches dataset}
\usage{
addMatchesColumns(matches)
}
\arguments{
\item{Matches}{Dataframe containing the matches from comparison_pool}
}
\value{
Dataframe containing top matches
}
\description{
Create additional columns for the matches dataset
}
\keyword{internal}
