% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_matches}
\alias{create_matches}
\title{Algorithm to find best matches from the comparison pool}
\usage{
create_matches(trimmed_pool, tm, num_matches = 3, replacement = TRUE)
}
\arguments{
\item{trimmed_pool}{Dataframe containing the pool from which matches
should be found}

\item{tm}{See rollmatch()}

\item{num_matches}{See rollmatch()}

\item{replacement}{See rollmatch()}
}
\value{
Dataframe containing top matches
}
\description{
Algorithm to find best matches from the comparison pool
}
\examples{
\donttest{ 
print('See add_balance_table for full example')
}

}
\keyword{internal}
