% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks_one}
\alias{run_checks_one}
\title{Run checks on variable inputs}
\usage{
run_checks_one(data, treat, tm, entry, id)
}
\arguments{
\item{data}{See rollmatch()}

\item{treat}{See rollmatch()}

\item{tm}{See rollmatch()}

\item{entry}{See rollmatch()}

\item{id}{See rollmatch()}
}
\description{
Run checks on variable inputs
}
\keyword{internal}
