% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks_two}
\alias{run_checks_two}
\title{Run checks on variable inputs}
\usage{
run_checks_two(data, alpha, standard_deviation, num_matches, replacement)
}
\arguments{
\item{data}{See rollmatch()}

\item{alpha}{See rollmatch()}

\item{standard_deviation}{See rollmatch()}

\item{num_matches}{See rollmatch()}

\item{replacement}{See rollmatch()}
}
\description{
Run checks on variable inputs
}
\keyword{internal}
