% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{runModel}
\alias{runModel}
\title{Run a logistic or probit model}
\usage{
runModel(model_type, match_on, reduced_data, id, treat, entry, tm, fm)
}
\arguments{
\item{model_type}{Use logistic regression ("logistic") or "probit"
regression ("probit") to estimate the predicted probability of participating}

\item{match_on}{Match on estimated propensity score ("pscore") or logit of
estimated propensity score ("logit").}

\item{reduced_data}{Dataframe of reduced treament and comparison data}

\item{id}{ID variable in dataset}

\item{treat}{Treatment variable in dataset}

\item{entry}{Entry quarter variable in dataset}
}
\value{
list of the model ($pred_model) and the resulting dataframe with
predicting values ($lr_result)
}
\description{
Run a logistic or probit model
}
\keyword{internal}
