% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trim_pool}
\alias{trim_pool}
\title{Use a caliper to trim the data to only observations within threshold}
\usage{
trim_pool(alpha, comparison_pool, scored_data, treat, tm,
  standard_deviation = "average")
}
\arguments{
\item{alpha}{See rollmatch()}

\item{comparison_pool}{Dataframe of comparison data to be trimmed from
compare_pool()}

\item{scored_data}{Dataframe of results from score_data()}

\item{treat}{See rollmatch()}

\item{tm}{See rollmatch()}

\item{standard_deviaion}{See rollmatch()

\dontrun{ 
print('See add_balance_table for full example')
}}
}
\value{
Dataframe of the trimmed comparisons based on the alpha value
}
\description{
Use a caliper to trim the data to only observations within threshold
}
\keyword{internal}
