% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trimPool}
\alias{trimPool}
\title{Caliper to trim the comparison data to only observations within threshold}
\usage{
trimPool(caliper, data_pool, lr_result, weighted_pooled_stdev = FALSE)
}
\arguments{
\item{caliper}{The pre-specified distance within which to allow matching.
The caliper width is calculated as the \code{caliper} multiplied by the
pooled standard deviation of the propensity scores or the logit of the
propensity scores - depending on the value of \code{match_on}.}

\item{lr_result}{Dataframe of results from model in runModel.}

\item{weighted_pooled_stdev}{Boolean. FALSE for average pooled standard
deviation and TRUE for weighted pooled standard deviation.}

\item{dataPool}{Dataframe of comparison data to be trimmed.}
}
\value{
Dataframe of the trimmed comparisons based on the caliper value
}
\description{
Caliper to trim the comparison data to only observations within threshold
}
\keyword{internal}
